added task specific postprocessing class to data refiner module.
